requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.36.1
python-dotenv==1.0.1
bcrypt==4.1.2
Werkzeug==3.0.3
docker==7.0.0
GitPython==3.1.43
SQLAlchemy==2.0.34
gunicorn==21.2.0
